Mover proyecto a la raíz y agregar métodos de Usuario

+++ src/main/java/pe/edu/upc/demoeva/repositories/UsuarioRepository.java
@@ -5,4 +5,4 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.demoeva.entities.Usuario;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> { }
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> { }
